refactor: use mpath.set instead of custom setPath
mpath has one shortcoming; when you try to set a path that doesn't
exists it silently ignores it.

In this commit I'm patching the `mpath.set` function by compaing the
result to the input value. If they are not the same it will try using
the parent path. The omitted part is used as the key for the value in an
object that becomes the new value.

Closes #69