refactor: reveret set path function
The mpath.set[path] function still had issues as demonstrated by
@AlbertHambardzumyan in #73

Not all of the implementation is reverted. This approach will ovrride
the mpath.set function and use our own setPath in its place.

If someone has the time to figure out how to get mpath.set to work place
create a PR.